refactor: handleInputChange 를 formHook 내부로 이관
